requirements.txt
requests>=2.9.1
paramiko>=1.16.0
beautifulsoup4>=4.4.1
pysnmp>=4.3.2
gnureadline>=6.3.3
python-nmap
